Locate the n2khab_data directory

Returns the absolute path of the n2khab_data directory by searching upwards in the directory hierarchy (starting in the working directory); the first hit will be returned.

locate_n2khab_data()

Details

If the n2khab_data_path option (or environment variable N2KHAB_DATA_PATH) is set, that value will be returned instead (see n2khab_options()).

See the data management advice in the vignette on data storage (run vignette("v020_datastorage")) for more information.

See also

Examples

if (FALSE) {
locate_n2khab_data()
}