n2khab: Providing preprocessed reference data for Flemish Natura 2000 habitat analyses

The n2khab package is an R package with preprocessing functions and standard reference data, useful for analyses regarding Flemish Natura 2000 habitats and regionally important biotopes (RIBs).

Details

See the Index for vignettes and for documentation on functions and datasets delivered with the package.

For contributing, see the README on GitHub.

See also