Documentation of included data source 'namelist'

'namelist' is a data source in the vc-format which provides names and (optionally) shortnames for IDs/codes used in other data sources. Multiple languages are supported.

Format

A vc-formatted data source. As such, it corresponds to a data frame with many rows and 4 variables:

code

A code used elsewhere.

lang

An IETF BCP 47 language tag, such as "en" or "nl", to identify the language of name and shortname.

name

The name corresponding to code and lang.

shortname

Optionally, a shorter variant of name.

Source

The 'namelist' data source has got its contents from the sources of several other n2khab-referencelists (see the source of those).

Typical way of loading

read_namelist()

read_namelist(lang = "nl")

Corresponding datafiles in the installed package

textdata/namelist.tsv

textdata/namelist.yml

See also