Return one of the available geospatial data sources representing administrative areas

Return one of the available geospatial data sources representing administrative areas

read_admin_areas(
  path = fileman_up("n2khab_data"),
  file = c("10_raw/flanders", "10_raw/provinces", "10_raw/sac"),
  dsn = c("flanders", "provinces", "sac")
)

Arguments

path
Location of the file. Considering the default value of the file argument, this should be the location of the folder 'n2khab_data'. By default, the first n2khab_data folder is used that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

file
The filename of the data source. May include a path prefix. The default follows the data management advice in the vignette on data storage.

dsn
A string, conforming to one of the data source names listed under Usage. Considering the default values of the path and file arguments, it is recommended to just use this argument. Defaults to "flanders".

Different data sources are handled differently by the function. The dsn argument states which data source is requested. It is also used to determine file if that argument is not specified, in order to select the right data file(s) from the n2khab_data folder.

Value

A Simple feature collection of geometry type MULTIPOLYGON or POLYGON.

Details

See section Usage to see which N2KHAB data sources are available. You get a list of the data source names and related information with XXXXXXXXX. You are referred to the raw N2KHAB-data collection at Zenodo to learn more about these data sources.

Examples

if (FALSE) {
flanders <- read_admin_areas(dsn = "flanders")
provinces <- read_admin_areas(dsn = "provinces")
sac <- read_admin_areas(dsn = "sac")
}