Return the 'env_pressures' data source as a tibble with human-readable attributes

Returns the included data source env_pressures as a tibble. Names, shortnames and explanations from namelist are added, in English by default.

read_env_pressures(
  path = pkgdatasource_path("textdata/env_pressures", ".tsv"),
  file = "env_pressures",
  file_namelist = "namelist",
  lang = "en"
)

Arguments

path
Location of the data sources env_pressures and namelist. The default is to use the location of the data sources as delivered by the installed package.

file
The filename of the env_pressures data source, without extension. The default is to use the file delivered by the installed package.

file_namelist
The filename of the namelist data source, without extension. The default is to use the file delivered by the installed package.

lang
An IETF BCP 47 language tag, such as "en" or "nl", to specify the language of human-readable attributes to be returned in the tibble.

Value

The env_pressures dataframe as a tibble, with human-readable text added for environmental pressures, pressure-classes and textual explanations according to the lang argument. The tibble has 35 rows and 6 variables. See env_pressures for documentation of the data-source's contents. See namelist for the link between codes or other identifiers and the corresponding text.

The human-readable attributes are represented by the following variables:

ep_abbrev

A (language-dependent) abbreviation (alternative code) of the environmental pressure. Is a factor with the level order coinciding with that of ep_code.

ep_name

The name of the environmental pressure. Is a factor with the level order coinciding with that of ep_code.

ep_class_name

The name of the environmental pressure's class. Is a factor with the level order coinciding with that of ep_class.

explanation

An explanation of the environmental pressure. Beware that this explanation is often shared between multiple environmental pressures! Hence the added explanation may cover more than is revealed by the environmental pressure's name.

Details

env_pressures is a data source in the vc-format which provides a checklist of environmental pressures, represented by codes, together with the pressure-class and the textual explanation.

read_env_pressures() reads the env_pressures data source, adds human-readable attributes and returns it as a tibble. A tibble is a dataframe that makes working in the tidyverse a little easier. By default, the data version delivered with the package is used and English text (lang = "en") is returned for names of environmental pressures and pressure-classes, and for textual explanations.

Recommended usage

read_env_pressures()

read_env_pressures(lang = "nl")

See also

Examples

read_env_pressures()

#> # A tibble: 35 x 6
#>    ep_code ep_abbrev   ep_name               ep_class ep_class_name  explanation
#>    <fct>   <fct>       <fct>                 <fct>    <fct>          <chr>      
#>  1 ep_011  011_struct  11 Change of physica… ep_clas… 1 Physical mo… <NA>       
#>  2 ep_012  012_soildy… 12 Soil dynamics inc… ep_clas… 1 Physical mo… <NA>       
#>  3 ep_013  013_soildy… 13 Soil dynamics dec… ep_clas… 1 Physical mo… <NA>       
#>  4 ep_014  014_aqconn  14 Aquatic connectiv… ep_clas… 1 Physical mo… <NA>       
#>  5 ep_015  015_terrco… 15 Terrestrial conne… ep_clas… 1 Physical mo… <NA>       
#>  6 ep_03.1 03.1_eutr_… 3.1 Eutrophication t… ep_clas… 3 Eutrophicat… <NA>       
#>  7 ep_03.2 03.2_eutr_… 3.2 Eutrophication t… ep_clas… 3 Eutrophicat… <NA>       
#>  8 ep_03.3 03.3_eutr_… 3.3 Eutrophication t… ep_clas… 3 Eutrophicat… <NA>       
#>  9 ep_03.4 03.4_eutr_… 3.4 Eutrophication t… ep_clas… 3 Eutrophicat… <NA>       
#> 10 ep_04.1 04.1_acidi… 4.1 Acidification th… ep_clas… 4 Acidificati… <NA>       
#> # … with 25 more rows

read_env_pressures(lang = "nl")

#> # A tibble: 35 x 6
#>    ep_code ep_abbrev  ep_name          ep_class ep_class_name explanation       
#>    <fct>   <fct>      <fct>            <fct>    <fct>         <chr>             
#>  1 ep_011  011_struct 11 Aanpassing v… ep_clas… 1 Ruimtebesl… "Aard van het sub…
#>  2 ep_012  012_boddy… 12 Toename bode… ep_clas… 1 Ruimtebesl… "Een toename resp…
#>  3 ep_013  013_boddy… 13 Afname bodem… ep_clas… 1 Ruimtebesl… "Een toename resp…
#>  4 ep_014  014_aqconn 14 Verlies van … ep_clas… 1 Ruimtebesl… "Barrières en iso…
#>  5 ep_015  015_terrc… 15 Verlies van … ep_clas… 1 Ruimtebesl… "Barrières en iso…
#>  6 ep_03.1 03.1_eutr… 3.1 Eutrofiërin… ep_clas… 3 Eutrofiëri… "Voor de levensge…
#>  7 ep_03.2 03.2_eutr… 3.2 Eutrofiërin… ep_clas… 3 Eutrofiëri… "Voor de levensge…
#>  8 ep_03.3 03.3_eutr… 3.3 Eutrofiërin… ep_clas… 3 Eutrofiëri… "Voor de levensge…
#>  9 ep_03.4 03.4_eutr… 3.4 Eutrofiërin… ep_clas… 3 Eutrofiëri… "Voor de levensge…
#> 10 ep_04.1 04.1_verz… 4.1 Verzuring v… ep_clas… 4 Verzuring   "Daling van de pH…
#> # … with 25 more rows