Return the data source habitatmap as an sf multipolygon layer

Returns the raw data source habitatmap (De Saeger et al., 2018) as a standardized sf multipolygon layer (tidyverse-styled, internationalized) in the Belgian Lambert 72 CRS (EPSG-code 31370). Given the size of the data source, this function takes a bit longer than usual to run.

read_habitatmap(
  path = fileman_up("n2khab_data"),
  file = "10_raw/habitatmap",
  select_hab = FALSE
)

Arguments

path
Location of the file. Considering the default value of the file argument, this should be the location of the folder 'n2khab_data'. By default, the first n2khab_data folder is used that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

file
The filename of the data source. May include a path prefix. The default follows the data management advice in the vignette on data storage (run vignette("v020_datastorage")).

select_hab
If TRUE only polygons that (partially) contain habitat or a regionally important biotope (RIB) are returned. The default value is FALSE.

Value

A Simple feature collection of type MULTIPOLYGON.

References

De Saeger S., Guelinckx R., Oosterlynck P., De Bruyn A., Debusschere K., Dhaluin P., Erens R., Hendrickx P., Hendrix R., Hennebel D., et al. (2018). Biologische Waarderingskaart en Natura 2000 Habitatkaart: Uitgave 2018. Rapporten van het Instituut voor Natuur- en Bosonderzoek. Instituut voor Natuur- en Bosonderzoek (INBO). DOI: https://doi.org/10.21436/inbor.15138099.

See also

Examples

if (FALSE) {
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'habitatmap'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

hm <- read_habitatmap()
hm
}