Return the data source habitatmap_stdized as a list of two objects

read_habitatmap_stdized returns the data source habitatmap_stdized as a list of two objects:

habitatmap_polygons: an sf object in the Belgian Lambert 72 CRS (EPSG-code 31370) with all polygons of the habitatmap that contain habitat or a regionally important biotope (RIB).

habitatmap_types: a tibble with information on the habitat and RIB types (HAB1, HAB2,..., HAB5) that occur within each polygon of habitatmap_polygons.

read_habitatmap_stdized(
  path = fileman_up("n2khab_data"),
  file = "20_processed/habitatmap_stdized/habitatmap_stdized.gpkg",
  version = "habitatmap_stdized_2018_v2"
)

Arguments

path
Location of the file. Considering the default value of the file argument, this should be the location of the folder 'n2khab_data'. By default, the first n2khab_data folder is used that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

file
The filename of the data source. May include a path prefix. The default follows the data management advice in the vignette on data storage (run vignette("v020_datastorage")).

version
Version ID of the data source. Defaults to the latest available version defined by the package.

Value

A list of two objects:

habitatmap_polygons: an sf object of habitatmap polygons with two attribute variables

polygon_id

description_orig: polygon description based on the orginal vegetation codes in the raw habitatmap

habitatmap_types: a tibble with following variables

polygon_id

code_orig: original vegetation code in raw habitatmap.

phab: proportion of polygon covered by type, as a percentage.

certain: TRUE when vegetation type is certain and FALSE when vegetation type is uncertain.

type: habitat or RIB type listed in types.

Details

The data source habitatmap_stdized is the processed version of the raw data source habitatmap (De Saeger et al., 2018). Every polygon in the habitatmap can consist of maximum 5 different vegetation types. This information is stored in the columns 'HAB1', HAB2',..., 'HAB5' of the attribute table. The fraction of each vegetation type within the polygons is stored in the columns 'PHAB1', 'PHAB2', ..., 'PHAB5'.

The data source habitatmap_stdized is a GeoPackage, available at Zenodo, that contains:

habitatmap_polygons: a spatial layer with every habitatmap polygon that contains a habitat or RIB type listed in types.

habitatmap_types: a table with the types that occur in each polygon.

The processing of the habitatmap_types tibble included following steps:

For some polygons the vegetation type is uncertain, and the vegetation code in the raw habitatmap data source consists of 2 or 3 possible types, separated with a ','. The different possible vegetation types are split up and one row is created for each of them. The variable certain will be FALSE if the original code consists of 2 or 3 possible vegetation types, and TRUE if only one vegetation type is provided.

For some polygons the original vegetation code in the habitatmap was not consistent with general coding syntax or with the type codes from the types. In that case the code was adjusted.

The R-code for creating the habitatmap_stdized data source can be found in the n2khab-preprocessing repository.

References

De Saeger S., Guelinckx R., Oosterlynck P., De Bruyn A., Debusschere K., Dhaluin P., Erens R., Hendrickx P., Hendrix R., Hennebel D., et al. (2018). Biologische Waarderingskaart en Natura 2000 Habitatkaart: Uitgave 2018. Rapporten van het Instituut voor Natuur- en Bosonderzoek. Instituut voor Natuur- en Bosonderzoek (INBO). DOI: https://doi.org/10.21436/inbor.15138099.

See also

Examples

if (FALSE) {
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'habitatmap_stdized'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

hms <- read_habitatmap_stdized()
hms_polygons <- hms$habitatmap_polygons
hms_types <- hms$habitatmap_types
hms_polygons
hms_types
}