Return the data source habitatmap_terr as a list of two objects

read_habitatmap_terr() returns the data source habitatmap_terr as a list of two objects: habitatmap_terr_polygons and habitatmap_terr_types. habitatmap_terr is the further interpreted, terrestrial part of habitatmap_stdized (see read_habitatmap_stdized).

read_habitatmap_terr(
  path = fileman_up("n2khab_data"),
  file = "20_processed/habitatmap_terr/habitatmap_terr.gpkg",
  keep_aq_types = TRUE,
  version = "habitatmap_terr_2018_v2"
)

Arguments

path
Location of the file. Considering the default value of the file argument, this should be the location of the folder 'n2khab_data'. By default, the first n2khab_data folder is used that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

file
The filename of the data source. May include a path prefix. The default follows the data management advice in the vignette on data storage.

keep_aq_types
Logical; TRUE by default. The data source habitatmap_terr aims at delineating all polygons with at least one (semi-)terrestrial type. For those polygons, it returns all known habitat types and RIBs as types. Hence, in several cases polygons do occur with a combination of terrestrial and aquatic types (see Details for a definition of 'aquatic'). Setting keep_aq_types = FALSE is convenient for use cases where one only wants to look at the (semi-)terrestrial types: this setting will discard all aquatic types in 'mixed' aquatic/terrestrial polygons. As each polygon always has at least one (semi-)terrestrial type, this will not affect the number of polygons returned, only the number of types.

version
Version ID of the data source. Defaults to the latest available version defined by the package.

Value

A list of two objects:

habitatmap_terr_polygons: a Simple feature collection of geometry type MULTIPOLYGON with four attribute variables:

polygon_id

description_orig: polygon description based on the original vegetation codes in the habitatmap data source

description: based on description_orig but with the interpreted type codes

source: states where description comes from: either habitatmap_stdized or habitatmap_stdized + interpretation

habitatmap_terr_types: a tibble with the following variables (the first 4 being identical to those in habitatmap_stdized):

polygon_id

code_orig

phab

certain

type: the interpreted habitat or RIB type

source: states where type comes from: either habitatmap_stdized or habitatmap_stdized + interpretation

Details

habitatmap_terr was derived from habitatmap_stdized as follows:

it excludes all polygons that are most probably aquatic habitat or RIB. These are the polygons for which all habitat or RIB types are aquatic. In the process, a distinction was also made between 2190_a and 2190_overig. There is no exclusion of aquatic types when these coexist with terrestrial types in the same polygon. The aquatic types are the types for which tag_2 == "HC3" in the types data source (tag_2 is the hydrological class; cf. the output of read_types());

it excludes types which most probably are no habitat or RIB at all. Those are the types where code_orig contains "bos" or is equal to "6510,gh" or "9120,gh";

it translates several main type codes into a corresponding subtype which they almost always represent: 6410 -> 6410_mo, 6430 -> 6430_hf, 6510 -> 6510_hu, 7140 -> 7140_meso, 9130 -> 9130_end;

it distinguishes types rbbhfl and rbbhf.

The data source habitatmap_terr is a GeoPackage, available at Zenodo, that contains:

habitatmap_terr_polygons: a spatial polygon layer

habitatmap_terr_types: a table with the types that occur in each polygon.

The R-code for creating the habitatmap_terr data source can be found in the n2khab-preprocessing repository.

See also

Examples

if (FALSE) {
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'habitatmap_terr'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

habmap_terr <- read_habitatmap_terr()
habmap_terr$habitatmap_terr_polygons
habmap_terr$habitatmap_terr_types

habmap_terr_noaq <- read_habitatmap_terr(keep_aq_types = FALSE)
habmap_terr_noaq$habitatmap_terr_polygons
habmap_terr_noaq$habitatmap_terr_types
}