Return the data source habitatstreams as an sf linestring layer or as a list

Returns the raw data source habitatstreams (Leyssen et al., 2020) as an sf linestring layer or as a list of two objects: the sf object (CRS: Belgian Lambert 72 (EPSG-code 31370)) plus a data frame with textual explanation of the values of the source_id variable.

read_habitatstreams(
  file = file.path(locate_n2khab_data(), "10_raw/habitatstreams"),
  source_text = FALSE
)

Arguments

file
The absolute or relative file path of the data source. The default follows the data management advice in the vignette on data storage (run vignette("v020_datastorage")). It uses the first n2khab_data folder that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

source_text
Logical, defaults to FALSE. If TRUE, a list is returned (see Value).

Value

With source_text = FALSE (default): a Simple feature collection of type LINESTRING.

With source_text = TRUE: a list of two objects:

lines: the same sf object as with source_text = FALSE.

sources: textual explanation on the values of the source_id variable in the sf object.

References

Leyssen A., Smeekens V., Denys L. (2020). Indicatieve situering van het Natura 2000 habitattype 3260. Submontane en laaglandrivieren met vegetaties behorend tot het Ranunculion fluitantis en het Callitricho-Batrachion. Uitgave 2020 (versie 1.7). Rapporten van het Instituut voor Natuur- en Bosonderzoek 2020 (34). Research Institute for Nature and Forest, Brussels. doi:10.21436/inbor.18903609 .

Examples

if (FALSE) {
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'habitatstreams'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

library(magrittr)
library(sf)
hs <- read_habitatstreams()
hs
hs2 <- read_habitatstreams(source_text = TRUE)
hs2
all.equal(
  hs %>% st_drop_geometry(),
  hs2$lines %>% st_drop_geometry()
)
}