Return the 'namelist' data source as a tibble

Returns the included data source namelist as a tibble, by default filtered according to English names and shortnames.

read_namelist(
  path = pkgdatasource_path("textdata/namelist", ".tsv"),
  file = "namelist",
  lang = "en"
)

Arguments

path
Location of the data source. The default is to use the location of the data source as delivered by the installed package.

file
The filename of the data source, without extension. The default is to use the file delivered by the installed package.

lang
An IETF BCP 47 language tag, such as "en" or "nl", to specify the language of name and shortname to be returned in the tibble. If lang = "all", the full namelist tibble is returned, i.e. containing all languages.

Value

The namelist dataframe as a tibble, filtered according to the lang argument. See namelist for documentation of the tibble's contents.

Details

namelist is a data source in the vc-format which provides names and (optionally) shortnames for IDs/codes used in other data sources.

read_namelist() reads it and returns it as a tibble. A tibble is a dataframe that makes working in the tidyverse a little easier. By default, the data version delivered with the package is used and only English names (lang = "en") are returned.

Recommended usage

read_namelist()

read_namelist(lang = "nl")

See also

Examples

read_namelist()

#> # A tibble: 282 x 4
#>    code    lang  name                                    shortname              
#>    <chr>   <chr> <chr>                                   <chr>                  
#>  1 1130    en    Estuaries                               Estuaries              
#>  2 1140    en    Mudflats and sandflats not covered by … Mud- and sandflats     
#>  3 1310    en    Salicornia and other annuals colonizin… Brackish pioneer habit…
#>  4 1310_p… en    Salicornia habitats dike inward         Salicornia habitats di…
#>  5 1310_zk en    Low saltmarshes Salicornia habitats di… Low saltmarshes dike o…
#>  6 1310_zv en    High saltmarshes Saginion habitats dik… High saltmarshes dike …
#>  7 1320    en    Spartina swards (Spartinion maritimae)  Spartina swards        
#>  8 1330    en    Atlantic salt meadows (Glauco-Puccinel… Atlantic salt meadows  
#>  9 1330_da en    Saltmarshes dike outward                Saltmarshes dike outwa…
#> 10 1330_h… en    Halophytic grasslands dike inward       Halophytic grassland   
#> # … with 272 more rows

read_namelist(lang = "nl")

#> # A tibble: 295 x 4
#>    code    lang  name                                       shortname           
#>    <chr>   <chr> <chr>                                      <chr>               
#>  1 1130    nl    Estuaria                                   estuaria            
#>  2 1140    nl    Bij eb droogvallende slikwadden en zandpl… bij eb droogvallend…
#>  3 1310    nl    Eenjarige pioniersvegetaties van slik- en… zilte pionierbegroe…
#>  4 1310_p… nl    binnendijks gelegen zeekraalvegetaties     binnendijkse zeekra…
#>  5 1310_zk nl    buitendijks laag schor met zeekraalvegeta… buitendijks laag sc…
#>  6 1310_zv nl    buitendijks hoog schor met zeevetmuurvege… buitendijks hoog sc…
#>  7 1320    nl    Schorren met slijkgrasvegetatie (Spartini… schorren met slijkg…
#>  8 1330    nl    Atlantische schorren (Glauco-Puccinelliet… Atlantische schorren
#>  9 1330_da nl    buitendijkse schorren                      buitendijkse schor  
#> 10 1330_h… nl    binnendijkse zilte vegetaties: zilte gras… zilte graslanden    
#> # … with 285 more rows