Return the data source shallowgroundwater as an sf multipolygon layer

Returns the raw data source shallowgroundwater as an sf multipolygon layer. The coordinate reference system is 'BD72 / Belgian Lambert 72' (EPSG-code 31370).

read_shallowgroundwater(
  file = file.path(locate_n2khab_data(),
    "10_raw/shallowgroundwater/shallowgroundwater.gpkg")
)

Arguments

file
The absolute or relative file path of the data source. The default follows the data management advice in the vignette on data storage (run vignette("v020_datastorage")). It uses the first n2khab_data folder that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

Value

A Simple feature collection of geometry type MULTIPOLYGON.

All attribute variables are logical variables referring to a data source (subset) or procedure that has contributed to the shallowgroundwater data source. If a variable is TRUE for a multipolygon, then the related data source (subset) or procedure has contributed to the multipolygon. For one multipolygon, several variables can be TRUE at the same time. Each combination of values occurs in only one multipolygon.

The attribute variables listed below are described in more detail at Zenodo; their description typically refers to the meaning of TRUE:

geomorph_wcoast: mainly concerns dune slacks, mud flats and salt marshes. Derived from: Cosyns et al. (2019)

anthrop_gwdep: zones located within a 100 m buffer around (almost) everywhere groundwater dependent habitat types (or regionally important biotopes) ánd situated within zones classified as 'anthropogenic' areas within the soil map. Derived from: soilmap_simple and habitatmap_terr data sources

narrowanthrop_gwdep: narrow zones classified as 'anthropogenic' areas within the soil map that include (almost) everywhere groundwater dependent habitat types (or regionally important biotopes). Derived from: soilmap_simple and habitatmap_terr data sources

drainage: soils that are at least moderately gleyic or wet. Derived from: soilmap_simple data source

dunes_gwdep: zones located within a 100 m buffer around (almost) everywhere groundwater dependent habitat types (or regionally important biotopes) ánd situated within zones classified as 'dunes' areas within the soil map. Derived from: soilmap_simple and habitatmap_terr data sources

peat_profile: variant of the soil profile indicates a superficial peaty cover, mostly on gleyic or permanently water saturated soil with or without profile development ((v)), eventually combined with strong anthropogenic influence ((o)). Derived from: soilmap_simple data source

peat_substr: soil substrate (layer underlying superficial layer, and lithologically diverging from it) consists of peat material starting at small (less than 75 cm; v) or moderate depths (75-125 cm; (v)), or a combination of the previous (v-). Derived from: soilmap_simple data source

peat_parentmat: parent material contains a mixture of at least 30% of peaty material. Derived from: soilmap_simple data source

peat_texture: soil consists of plain peat material. Derived from: soilmap_simple data source

phys_system: polygons designated as seepage areas where groundwater is supposed to gather after having infiltrated elsewhere (infiltration areas) and being transported through the landscape (passage areas). Derived from: Lhermitte & Honnay (1994)

zwin: the contour of the Zwin Nature Reserve in the most eastern part of the Flemish coastal area. Derived from: Open Street Map, consulted 2021-10-06 by QGIS plugin QuickOSM

habitat_1130: polygons located within estuaries (habitat type 1130). Derived from: habitatmap data source

gwdepth_coast: locations with estimated average lowest groundwater table less than 2.5 m below soil surface (shallow groundwater), based on interpolation of measured groundwater levels in areas along the Flemish coast with sufficient gauge densities. Derived from: the Watina database of the Research Institute for Nature and Forest (INBO)

gwdepth_local: mean lowest groundwater level less than 2 m below soil surface (MLW) in a large military training site for which no information is available in the soil map of Flanders. Derived from: Batelaan et al. (2012)

seepage: area with modelled seepage fluxes of at least 0.8 mm/day in the central part of eastern Flanders. Derived from: Batelaan & De Smedt (1994)

peat_survey: local peaty zones evaluated by simple measurements of the depth of plain and superficial peaty soil layers at regular intervals (about 20 m). Derived from: local inventories executed or compiled by the Research Institute for Nature and Forest (INBO)

duneslack: polygons with dune slack vegetations along the Flemish coast that typically imply shallow groundwater levels. Derived from: Provoost et al. (2020)

Details

The data source shallowgroundwater represents the areas in the Flemish region of Belgium where the mean lowest groundwater level (MLW; Knotters & Van Walsum, 1997; Van Heesen, 1970) is estimated to be less than approximately 2 m below soil surface (hence, 'shallow' groundwater). Groundwater dependent species and communities can be expected to be present mostly within these areas.

The data source is a GeoPackage, available at Zenodo, and contains a single spatial multipolygon layer shallowgroundwater.

The data source has been compiled from subsets of various other spatial data sources. This process is described in more detail at Zenodo. The R code for finishing the shallowgroundwater data source, starting from an intermediate result, can be found in the n2khab-preprocessing repository.

References

Batelaan O., De Becker P., El-Rawy M., Herr C., Schneidewind, U. (2012). Doorrekenen van maatregelen voor herstel van vochtige heidevegetaties op het Schietveld van Houthalen-Helchteren via grondwatermodellering. Vrije Universiteit Brussel/Instituut voor Natuur en Bosonderzoek.

Batelaan O., De Smedt F. (1994). Regionale grondwaterstroming rond een aantal kwelafhankelijke natuurgebieden. Vrije Universiteit Brussel.

Cosyns E., Bollengier B., Provoost S. (2019). Masterplan en juridische basis voor grensoverschrijdende samenwerking en bescherming als een transnationaal natuurpark van de duinen tussen Dunkerque (Frankrijk) en Westende (België). Partim Masterplan. Rapport in opdracht van Agentschap Natuur en Bos, Conservatoire de l'espace littoral et des rivages lacustres, Conseil Général Département du Nord. https://www.natuurenbos.be/sites/default/files/inserted-files/masterplan_flandre_ned20200210def.pdf

Knotters M. & van Walsum P.E.V. (1997). Estimating fluctuation quantities from time series of water-table depths using models with a stochastic component. Journal of Hydrology 197 (1): 25–46. doi:10.1016/S0022-1694(96)03278-7 .

Lhermitte K., Honnay O. (1994). Kartering van het Fysisch Systeem en de Ruimtelijke structuren in Vlaanderen op schaal 1 : 50 000. Stichting Plattelandsbeleid 1994, Boerenbond, Leuven.

Provoost S., Van Gompel W. & Vercruysse E. (2020). Beheerevaluatie kust. Eindrapport 2015-2019. Rapporten van het Instituut voor Natuur- en Bosonderzoek 2020 (18). Instituut voor Natuur- en Bosonderzoek, Brussel. doi:10.21436/inbor.18039583 .

Van Heesen H.C. (1970). Presentation of the seasonal fluctuation of the water table on soil maps. Geoderma 4 (3): 257–278. doi:10.1016/0016-7061(70)90006-6 .

See also

Examples

if (FALSE) {
shallowgroundwater <- read_shallowgroundwater()
shallowgroundwater
}