Return the data source watersurfaces as an sf polygon layer

Returns the raw data source watersurfaces (Packet et al., 2018) as a standardized sf polygon layer (tidyverse-styled, internationalized) in the Belgian Lambert 72 CRS (EPSG-code 31370).

read_watersurfaces(
  path = fileman_up("n2khab_data"),
  file = "10_raw/watersurfaces",
  extended = FALSE
)

Arguments

path
Location of the file. Considering the default value of the file argument, this should be the location of the folder 'n2khab_data'. By default, the first n2khab_data folder is used that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

file
The filename of the data source. May include a path prefix. The default follows the data management advice in the vignette on data storage (run vignette("v020_datastorage")).

extended
Logical. Should names or explanations of codes be added as extra variables in the result? Currently only applies to wfd_type; if TRUE, a variable wfd_type_name is added. Defaults to FALSE.

Value

A Simple feature collection of type POLYGON, sorted by polygon_id, with the following variables (not mentioning extra 'name' variables for extended = TRUE):

polygon_id: code of the polygon;

wfd_code: optional; Flemish code of the water body with regard to the Water Framework Directive (WFD);

hyla_code: optional; code of the watersurface according to the Flemish working group 'Hyla', a working group on amphibians & reptiles;

name: name of the watersurface;

area_name: name of the area;

wfd_type: type code according to the Flemish WFD typology (Denys, 2009);

wfd_type_certain: Logical. Is there high confidence about the wfd_type determination?

depth_class: class of water depth;

connectivity: connectivity class;

usage: usage class.

Details

See Packet et al. (2018) for an elaborate explanation of the data source and its attributes.

References

Denys L. (2009). Een a posteriori typologie van stilstaande, zoete wateren in Vlaanderen. Rapporten van het Instituut voor Natuur- en Bosonderzoek INBO.R.2009.34. Instituut voor Natuur- en Bosonderzoek, Brussel.

Packet J., Scheers K., Smeekens V., Leyssen A., Wils C. & Denys L. (2018). Watervlakken versie 1.0: polygonenkaart van stilstaand water in Vlaanderen. Een nieuw instrument voor onderzoek, water-, milieu- en natuurbeleid. Rapporten van het Instituut voor Natuur- en Bosonderzoek 2018 (14). Instituut voor Natuur- en Bosonderzoek, Brussel. DOI: https://doi.org/10.21436/inbor.14178464.

See also

Examples

if (FALSE) {
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the 'watersurfaces'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

ws <- read_watersurfaces()
ws
summary(ws)

ws2 <- read_watersurfaces(extended = TRUE)
ws2
}