Documentation of included data source 'scheme_types'

'scheme_types' is a data source in the vc-format which lists the types (using the type-code from types) that belong to each N2KHAB (monitoring or research) scheme (using the scheme-code from schemes). It also defines typegroup memberships of the types within specific schemes, if applicable. The codes of schemes and types (and optionally: typegroups) are explained in the data source namelist (which can accommodate multiple languages).

Format

A vc-formatted data source. As such, it corresponds to a dataframe with 3 variables:

scheme

Code of the scheme, as a factor, and explained by namelist. It must be represented in schemes.

type

Code of the type, as a factor, and explained by namelist. It must be represented in types.

typegroup

An optional code (and optionally explained by namelist), declaring the typegroup to which a type belongs within the specified scheme. Typegroups point out which types are considered together as a group in inferences for the scheme at hand.

Each combination of scheme and type must be unique.

Source

For the MNE-schemes:

the link between schemes and types comes from a vc-formatted, Dutch-language data source which can be found in the n2khab Github repository: misc/generate_textdata/rawraw_data/10_compmeetnet_types_milieudrukken. This data source was originally generated in the n2khab-mne-selections repository.

the declaration of typegroups comes from this googlesheet. Currently, the typegroups are kept up to date both in the googlesheet and in the data source.

For the MHQ-schemes, the list of sampled types is based on the report of Westra et al. (2014).

Typical way of loading

read_scheme_types()

read_scheme_types(lang = "nl")

Corresponding datafiles in the installed package

textdata/scheme_types.tsv

textdata/scheme_types.yml

See also