Documentation of included data source 'schemes'

'schemes' is a data source in the vc-format which provides a list of (monitoring) schemes for N2KHAB monitoring programmes or other N2KHAB projects, together with defining attributes and optional information. The codes of schemes, programmes, attributes and tags are explained in the data source namelist (which can accommodate multiple languages).

Format

A vc-formatted data source. As such, it corresponds to a dataframe with 10 variables:

scheme

Code of the scheme, as a factor. This is the ID for use in diverse workflows and datasets. Corresponding names and shortnames in multiple languages are to be found in namelist. Contains no duplicates! A 'scheme' refers to a monitoring or research setup that determines which types (habitat/RIBs) are to be investigated for a question or for a bunch of related questions.

programme

Code of the programme to which the scheme belongs. The programme's code is explained by namelist. Is a factor. A 'programme' refers to a N2KHAB monitoring programme or research project. One programme can correspond to multiple schemes. At least the monitoring programmes MHQ and MNE are present.

attribute_1

The first defining attribute of the scheme. Typically, the code is explained by namelist. Is a factor.

In MNE, this is used to declare the 'compartment superscheme' (scheme collection) to which the scheme belongs, and which is named after the environmental compartment (however note that the surfacewater superscheme comprises two environmental compartments).

In MHQ, this is used to define the target habitat type of the monitoring scheme.

attribute_2

A second defining attribute of the scheme (if needed). Typically, the code is explained by namelist. Is a factor.

In MNE, this provides the code of the environmental pressure to which the scheme is related. It must be represented in env_pressures.

attribute_3

A third defining attribute of the scheme (if needed). Typically, the code is explained by namelist. Is a factor.

In MNE, this is used to declare an optional partitioning of the scheme that would be constituted by attributes 1 and 2 alone (environmental compartment and pressure), which will hence divide the concerned types in at least two groups. This typically has to do with differences in the considered environmental variables or compatibility of interpretation. It is not meant to distinguish between different types or typegroups for inferences, so types are kept in the same scheme as long as the measured values have compatibility of interpretation. To declare which types or typegroups should be distinguished in inferences, use the typegroup variable in the scheme_types data source!

spatial_restriction

Optional further defining specification of a scheme (in English): spatial restrictions, superposed on the restrictions already generated by the attributes.

notes

Optional. In English. For additional notes on the scheme definition, if necessary.

tag_1

Optional tag, e.g. a categorization ID explained by namelist.

In MNE, this is used to tag which schemes are focal schemes and which ones are secondary schemes.

In MHQ, this is used to separate aquatic and terrestial monitoring schemes.

tag_2

Optional tag, e.g. a categorization ID explained by namelist.

tag_3

Optional tag, e.g. a categorization ID explained by namelist.

Source

For the MNE-schemes: a vc-formatted, Dutch-language data source which can be found in the n2khab Github repository: misc/generate_textdata/rawraw_data/10_compmeetnet_types_milieudrukken. This data source was originally generated in the n2khab-mne-selections repository (https://gitlab.com/florisvdh/n2khab-mne-selections).

More information on the MHQ-schemes can be found in Westra et al. (2014).

Typical way of loading

read_schemes()

read_schemes(lang = "nl")

Corresponding datafiles in the installed package

textdata/schemes.tsv

textdata/schemes.yml

See also