Documentation of included data source 'types'

'types' is a data source in the vc-format which provides a checklist of types, represented by their current codes, together with several attributes. A 'type' refers to either a (main) habitat type, a habitat subtype or a regionally important biotope (RIB). The codes of types, typeclasses and further attributes and tags are explained in the data source namelist (which can accommodate multiple languages).

Format

A vc-formatted data source. As such, it corresponds to a data frame with several variables:

type

Code of the type, as a factor. This is the ID for use in diverse workflows and datasets. Names in multiple languages are to be found in namelist. Only currently active codes are kept in this list, in order to avoid confusion (this especially relates to habitat subtypes and RIBs). Contains no duplicates!

typelevel

A factor that labels the type as either "main_type" or "subtype".

main_type

The main type that corresponds with type, as a factor. Each type is either a subtype of a main type, or is a main type itself. This is indicated by typelevel.

typeclass

A code explained by namelist, corresponding to the typeclass. Is a factor.

hydr_class

A code explained by namelist, corresponding to the hydrological class. Is a factor.

groundw_dep

A code explained by namelist, corresponding to the groundwater dependency category. Is a factor.

flood_dep

A code explained by namelist, corresponding to the flood dependency category. Is a factor. Note that flood dependency is only defined for (semi-)terrestrial types, hence for aquatic types (hydrological class HC3) it is NA.

tag_1

Optional tag, e.g. a categorization ID explained by namelist. Currently used to indicate subcategories within a few typeclasses, or to differentiate between lotic and lentic aquatic types.

tag_2

Optional tag, e.g. a categorization ID explained by namelist.

tag_3

Optional tag, e.g. a categorization ID explained by namelist.

Source

Most information comes from this googlesheet. Currently, the googlesheet and the data source are both kept up-to-date. However only the 'types' data source is under version control.

The source for the hydrological class attribute is a vc-formatted file stored in the package source code. It is read by the 'generate_textdata' bookdown project which generates the 'types' data source. The referred vc-formatted file was derived from a yet unpublished database on the interrelations between types, hydrological classes, environmental compartments and their characteristics, and environmental pressures.

The source for the groundwater and flood dependency attributes is this googlesheet. Currently, the googlesheet and the data source are both kept up-to-date. However only the 'types' data source is under version control.

Typical way of loading

read_types()

read_types(lang = "nl")

Corresponding datafiles in the installed package

textdata/types.csv

textdata/types.yml

See also