Documentation of included data source 'env_pressures'

'env_pressures' is a data source in the vc-format which provides a checklist of environmental pressures, represented by codes, together with the pressure-class and the textual explanation (with optional remarks). The codes of environmental pressures, pressure-classes and explanations are explained in the data source namelist (which can accommodate multiple languages).

Format

A vc-formatted data source. As such, it corresponds to a data frame with 35 rows and 3 variables:

ep_code

Code of the environmental pressure, as a factor. This is the ID for use in diverse workflows and datasets. Corresponding names and abbreviations in multiple languages are stored in namelist (as name and shortname, respectively). The abbreviation may be seen as an alternative, language-dependent code. Contains no duplicates!

ep_class

A code explained by namelist, corresponding to the environmental pressure's class. Is a factor.

explanation

A code explained by namelist, corresponding to the explanation on the environmental pressure, and optional remarks. Explanation and remarks are stored in namelist (as name and shortname, respectively)

Source

The latest worksheet of this googlesheet. Currently, the googlesheet and the data source are both kept up-to-date.

Typical way of loading

read_env_pressures()

read_env_pressures(lang = "nl")

Corresponding datafiles in the installed package

textdata/env_pressures.csv

textdata/env_pressures.yml

See also