n2khab: Providing Preprocessed Reference Data for Flemish Natura 2000 Habitat Analyses

The n2khab package is an R package with preprocessing functions and standard reference data, useful for analyses regarding Flemish Natura 2000 habitats and regionally important biotopes (RIBs).

Details

See the Index for vignettes and for documentation on functions and datasets delivered with the package.

For contributing, see the README on GitHub.

See also

Author

Maintainer: Floris Vanderhaeghe floris.vanderhaeghe@inbo.be (ORCID)

Authors:

Toon Westra toon.westra@inbo.be (ORCID)

Cécile Herr cecile.herr@inbo.be (ORCID)

Hans Van Calster hans.vancalster@inbo.be (ORCID)

Other contributors:

Dries Adriaens dries.adriaens@inbo.be (ORCID) [contributor]

Maud Raman maud.raman@inbo.be (ORCID) [contributor]

Jan Wouters jan.wouters@inbo.be (ORCID) [contributor]

Steven De Saeger steven.desaeger@inbo.be (ORCID) [contributor]

Els Lommelen els.lommelen@inbo.be (ORCID) [contributor]

Luc Denys luc.denys@inbo.be (ORCID) [contributor]

An Leyssen an.leyssen@inbo.be (ORCID) [contributor]

Patrik Oosterlynck patrik.oosterlynck@inbo.be (ORCID) [contributor]

Jeroen Vanden Borre jeroen.vandenborre@inbo.be (ORCID) [contributor]

Nathalie Cools nathalie.cools@inbo.be (ORCID) [contributor]

Bruno De Vos bruno.devos@inbo.be (ORCID) [contributor]

Suzanna Lettens suzanna.lettens@inbo.be (ORCID) [contributor]

Arne Verstraeten arne.verstraeten@inbo.be (ORCID) [contributor]

Research Institute for Nature and Forest info@inbo.be [copyright holder]