Return the 'env_pressures' data source as a tibble with human-readable attributes

Returns the included data source env_pressures as a tibble. Names, shortnames, explanations and optional remarks from namelist are added, in English by default.

read_env_pressures(
  path = pkgdatasource_path("textdata/env_pressures", ".yml"),
  file = "env_pressures",
  file_namelist = "namelist",
  lang = "en"
)

Arguments

path
Location of the data sources env_pressures and namelist. The default is to use the location of the data sources as delivered by the installed package.

file
The filename of the env_pressures data source, without extension. The default is to use the file delivered by the installed package.

file_namelist
The filename of the namelist data source, without extension. The default is to use the file delivered by the installed package.

lang
An IETF BCP 47 language tag, such as "en" or "nl", to specify the language of human-readable attributes to be returned in the tibble.

Value

The env_pressures data frame as a tibble, with human-readable text added for environmental pressures, pressure-classes and textual explanations and remarks according to the lang argument. The tibble has 35 rows and 7 variables. See env_pressures for documentation of the data-source's contents. See namelist for the link between codes or other identifiers and the corresponding text.

The human-readable attributes are represented by the following variables:

ep_abbrev

A (language-dependent) abbreviation (alternative code) of the environmental pressure. Is a factor with the level order coinciding with that of ep_code.

ep_name

The name of the environmental pressure. Is a factor with the level order coinciding with that of ep_code.

ep_class_name

The name of the environmental pressure's class. Is a factor with the level order coinciding with that of ep_class.

explanation

An explanation of the environmental pressure.

remarks

Optional remarks about the environmental pressure.

Details

env_pressures is a data source in the vc-format which provides a checklist of environmental pressures, represented by codes, together with the pressure-class and the textual explanation.

read_env_pressures() reads the env_pressures data source, adds human-readable attributes and returns it as a tibble. A tibble is a data frame that makes working in the tidyverse a little easier. By default, the data version delivered with the package is used and English text (lang = "en") is returned for names of environmental pressures and pressure-classes, and for textual explanations and remarks.

Recommended usage

read_env_pressures()

read_env_pressures(lang = "nl")

See also

Examples

read_env_pressures()
#> # A tibble: 35 × 7
#>    ep_code ep_abbrev        ep_name   ep_class ep_class_name explanation remarks
#>    <fct>   <fct>            <fct>     <fct>    <fct>         <chr>       <chr>  
#>  1 ep_011  011_struct       11 Chang… ep_clas… 1 Physical m… NA          NA     
#>  2 ep_012  012_soildyn_incr 12 Soil … ep_clas… 1 Physical m… NA          NA     
#>  3 ep_013  013_soildyn_decr 13 Soil … ep_clas… 1 Physical m… NA          NA     
#>  4 ep_014  014_aqconn       14 Aquat… ep_clas… 1 Physical m… NA          NA     
#>  5 ep_015  015_terrconn     15 Terre… ep_clas… 1 Physical m… NA          NA     
#>  6 ep_03.1 03.1_eutr_air    3.1 Eutr… ep_clas… 3 Eutrophica… NA          NA     
#>  7 ep_03.2 03.2_eutr_soil   3.2 Eutr… ep_clas… 3 Eutrophica… NA          NA     
#>  8 ep_03.3 03.3_eutr_gw     3.3 Eutr… ep_clas… 3 Eutrophica… NA          NA     
#>  9 ep_03.4 03.4_eutr_sw     3.4 Eutr… ep_clas… 3 Eutrophica… NA          NA     
#> 10 ep_04.1 04.1_acidif_air  4.1 Acid… ep_clas… 4 Acidificat… NA          NA     
#> # ℹ 25 more rows
read_env_pressures(lang = "nl")
#> # A tibble: 35 × 7
#>    ep_code ep_abbrev       ep_name    ep_class ep_class_name explanation remarks
#>    <fct>   <fct>           <fct>      <fct>    <fct>         <chr>       <chr>  
#>  1 ep_011  011_struct      11 Aanpas… ep_clas… 1 Ruimtebesl… Aard van h… NA     
#>  2 ep_012  012_boddyn_plus 12 Toenam… ep_clas… 1 Ruimtebesl… Een toenam… NA     
#>  3 ep_013  013_boddyn_min  13 Afname… ep_clas… 1 Ruimtebesl… Een afname… NA     
#>  4 ep_014  014_aqconn      14 Verlie… ep_clas… 1 Ruimtebesl… Barrières … NA     
#>  5 ep_015  015_terrconn    15 Verlie… ep_clas… 1 Ruimtebesl… Barrières … NA     
#>  6 ep_03.1 03.1_eutr_atm   3.1 Eutro… ep_clas… 3 Eutrofiëri… Voor de le… NA     
#>  7 ep_03.2 03.2_eutr_bod   3.2 Eutro… ep_clas… 3 Eutrofiëri… Voor de le… NA     
#>  8 ep_03.3 03.3_eutr_gw    3.3 Eutro… ep_clas… 3 Eutrofiëri… Voor de le… NA     
#>  9 ep_03.4 03.4_eutr_ow    3.4 Eutro… ep_clas… 3 Eutrofiëri… Voor de le… NA     
#> 10 ep_04.1 04.1_verzu_atm  4.1 Verzu… ep_clas… 4 Verzuring   Daling van… NA     
#> # ℹ 25 more rows