Return the data source habitatmap_stdized as a list of two objects

read_habitatmap_stdized returns the data source habitatmap_stdized as a list of two objects:

habitatmap_polygons: an sf object in the Belgian Lambert 72 CRS (EPSG-code 31370) with all polygons of the habitatmap that contain habitat or a regionally important biotope (RIB).

habitatmap_types: a tibble with information on the habitat and RIB types (HAB1, HAB2,..., HAB5) that occur within each polygon of habitatmap_polygons.

read_habitatmap_stdized(
  file = file.path(locate_n2khab_data(),
    "20_processed/habitatmap_stdized/habitatmap_stdized.gpkg"),
  version = c("habitatmap_stdized_2020_v1", "habitatmap_stdized_2018_v2",
    "habitatmap_stdized_2018_v1")
)

Arguments

file
The absolute or relative file path of the data source. The default follows the data management advice in the vignette on data storage (run vignette("v020_datastorage")). It uses the first n2khab_data folder that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

version
Version ID of the data source. Defaults to the latest available version defined by the package.

Value

A list of two objects:

habitatmap_polygons: an sf object of habitatmap polygons with two attribute variables

polygon_id

description_orig: polygon description based on the orginal type codes in the raw habitatmap

habitatmap_types: a tibble with following variables

polygon_id

type: habitat or RIB type listed in types.

certain: TRUE when the type is certain and FALSE when the type is uncertain.

code_orig: original type code in raw habitatmap.

phab: proportion of polygon covered by type, as a percentage.

Since version habitatmap_stdized_2020_v1, rows are unique only by the combination of the polygon_id, type and certain columns.

Details

The data source habitatmap_stdized is the processed version of the raw data source habitatmap (De Saeger et al., 2020). Every polygon in the habitatmap can consist of maximum 5 different types. This information is stored in the columns 'HAB1', HAB2',..., 'HAB5' of the attribute table. The fraction of each type within the polygons is stored in the columns 'PHAB1', 'PHAB2', ..., 'PHAB5'.

The data source habitatmap_stdized is a GeoPackage, available at Zenodo, that contains:

habitatmap_polygons: a spatial layer with every habitatmap polygon that contains a habitat or RIB type listed in types.

habitatmap_types: a table with the types that occur in each polygon.

The processing of the habitatmap_types tibble included following steps:

For some polygons the type is uncertain, and the type code in the raw habitatmap data source consists of 2 or 3 possible types, separated with a ','. The different possible types are split up and one row is created for each of them, with phab for each new row simply set to the original value of phab. The variable certain will be FALSE if the original code consists of 2 or 3 possible types, and TRUE if only one type is provided.

Some polygons contain both a standing water habitat type and rbbmr: 3130_rbbmr, 3140_rbbmr, 3150_rbbmr and 3160_rbbmr. Since habitatmap_stdized_2020_v1, the two types 31xx and rbbmr are split up and one row is created for each of them, with phab for each new row simply set to the original value of phab. The variable certain in this case will be TRUE for both types.

After those first two steps, a given polygon could contain the same type with the same value for certain repeated several times, e.g. when 31xx_rbbmr is present with phab = yy% and 31xx is present with phab = zz%. In that case the rows with the same polygon_id, type and certain were gathered into one row and the respective phab values were added up.

For some polygons the original type code in the habitatmap was not consistent with general coding syntax or with the type codes from the types data source. In that case the code was adjusted.

The R-code for creating the habitatmap_stdized data source can be found in the n2khab-preprocessing repository.

References

De Saeger, S., Guelinckx, R., Oosterlynck, P., De Bruyn, A., Debusschere, K., Dhaluin, P., Erens, R., Hendrickx, P., Hennebel, D., Jacobs, I., Kumpen, M., Opdebeeck, J., Spanhove, T., Tamsyn, W., Van Oost, F., Van Dam, G., Van Hove, M., Wils, C., Paelinckx, D. (2020). Biologische Waarderingskaart en Natura 2000 Habitatkaart, uitgave 2020. (Rapporten van het Instituut voor Natuur- en Bosonderzoek; Nr. 35). Instituut voor Natuur- en Bosonderzoek (INBO). doi:10.21436/inbor.18840851 .

De Saeger, S., Oosterlynck, P. & Paelinckx, D. (2017). The Biological Valuation Map (BVM): a field-driven survey of land cover and vegetation in the Flemish Region of Belgium. Documents phytosociologiques - Actes du colloque de Saint-Mandé 2012 - Prodrome et cartographie des végétations de France - 2017. Vol. 6: 372-382.

See also

Examples

if (FALSE) {
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'habitatmap_stdized'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

hms <- read_habitatmap_stdized()
hms_polygons <- hms$habitatmap_polygons
hms_types <- hms$habitatmap_types
hms_polygons
hms_types
}