Return the 'namelist' data source as a tibble

Returns the included data source namelist as a tibble, by default filtered according to English names and shortnames.

read_namelist(
  path = pkgdatasource_path("textdata/namelist", ".yml"),
  file = "namelist",
  lang = "en"
)

Arguments

path
Location of the data source. The default is to use the location of the data source as delivered by the installed package.

file
The filename of the data source, without extension. The default is to use the file delivered by the installed package.

lang
An IETF BCP 47 language tag, such as "en" or "nl", to specify the language of name and shortname to be returned in the tibble. If lang = "all", the full namelist tibble is returned, i.e. containing all languages.

Value

The namelist data frame as a tibble, filtered according to the lang argument. See namelist for documentation of the tibble's contents.

Details

namelist is a data source in the vc-format which provides names and (optionally) shortnames for IDs/codes used in other data sources.

read_namelist() reads it and returns it as a tibble. A tibble is a data frame that makes working in the tidyverse a little easier. By default, the data version delivered with the package is used and only English names (lang = "en") are returned.

Recommended usage

read_namelist()

read_namelist(lang = "nl")

See also

Examples

read_namelist()
#> # A tibble: 189 × 4
#>    code     lang  name                                                 shortname
#>    <chr>    <chr> <chr>                                                <chr>    
#>  1 1130     en    Estuaries                                            Estuaries
#>  2 1140     en    Mudflats and sandflats not covered by seawater at l… Mud- and…
#>  3 1310     en    Salicornia and other annuals colonizing mud and sand Brackish…
#>  4 1310_pol en    Salicornia habitats dike inward                      Salicorn…
#>  5 1310_zk  en    Low saltmarshes Salicornia habitats dike outward     Low salt…
#>  6 1310_zv  en    High saltmarshes Saginion habitats dike outward      High sal…
#>  7 1320     en    Spartina swards (Spartinion maritimae)               Spartina…
#>  8 1330     en    Atlantic salt meadows (Glauco-Puccinellietalia mari… Atlantic…
#>  9 1330_da  en    Saltmarshes dike outward                             Saltmars…
#> 10 1330_hpr en    Halophytic grasslands dike inward                    Halophyt…
#> # ℹ 179 more rows
read_namelist(lang = "nl")
#> # A tibble: 217 × 4
#>    code     lang  name                                                 shortname
#>    <chr>    <chr> <chr>                                                <chr>    
#>  1 1130     nl    Estuaria                                             estuaria 
#>  2 1140     nl    Bij eb droogvallende slikwadden en zandplaten        bij eb d…
#>  3 1310     nl    Eenjarige pioniervegetaties van slik- en zandgebied… zilte pi…
#>  4 1310_pol nl    Binnendijks gelegen zeekraalvegetaties               binnendi…
#>  5 1310_zk  nl    Buitendijks laag schor met zeekraalvegetaties        buitendi…
#>  6 1310_zv  nl    Buitendijks hoog schor met zeevetmuurvegetaties (Sa… buitendi…
#>  7 1320     nl    Schorren met slijkgrasvegetatie (Spartinion maritim… schorren…
#>  8 1330     nl    Atlantische schorren (Glauco-Puccinellietalia marit… Atlantis…
#>  9 1330_da  nl    Buitendijkse schorren                                buitendi…
#> 10 1330_hpr nl    Binnendijkse zilte vegetaties: zilte graslanden      zilte gr…
#> # ℹ 207 more rows