Return the data source watersurfaces_hab as a list of two objects

read_watersurfaces_hab returns the data source watersurfaces_hab as a list of two objects:

watersurfaces_polygons: an sf object in the Belgian Lambert 72 CRS (EPSG-code 31370) with all polygons that contain standing water types (habitat or RIB).

watersurfaces_types: a tibble with information on the standing water types (HAB1, HAB2,..., HAB5) that occur within each polygon of watersurfaces_polygons.

read_watersurfaces_hab(
  file = file.path(locate_n2khab_data(),
    "20_processed/watersurfaces_hab/watersurfaces_hab.gpkg"),
  interpreted = FALSE,
  version = c("watersurfaces_hab_v4", "watersurfaces_hab_v3", "watersurfaces_hab_v2",
    "watersurfaces_hab_v1")
)

Arguments

file
The absolute or relative file path of the data source. The default follows the data management advice in the vignette on data storage (run vignette("v020_datastorage")). It uses the first n2khab_data folder that is found when sequentially climbing up 0 to 10 levels in the file system hierarchy, starting from the working directory.

interpreted
If TRUE, the interpreted subtype is provided when the subtype is missing. This only applies to type 3130. When the subtype is missing for 3130, we interpret it as 3130_aom.

version
Version ID of the data source. Defaults to the latest available version defined by the package.

Value

A list of two objects:

watersurfaces_polygons: an sf object of standing water polygons with four attribute variables:

polygon_id

polygon_id_ws: id for the polygon in the watersurface map

polygon_id_habitatmap: id's of all overlapping polygons of habitatmap_stdized that contain standing water habitat. The different id's are separated by '+'.

description_orig: descriptions of all overlapping polygons of habitatmap_stdized that contain standing water habitat. The different descriptions are separated by '+'.

watersurfaces_types: a tibble with following variables:

polygon_id

type: habitat or RIB type listed in types.

certain: TRUE when the type is certain and FALSE when the type is uncertain.

code_orig: original type code in raw habitatmap.

Details

The data source watersurfaces_hab is a combination of habitatmap_stdized (see read_habitatmap_stdized) and the watersurface map of Flanders. It contains all standing water types in Flanders.

The data source watersurfaces_hab is a GeoPackage, available at Zenodo, that contains:

watersurfaces_hab_polygons: a spatial layer with all polygons that contain standing water types listed in types.

watersurfaces_hab_types: a table in which every row corresponds with a combination of polygon and type.

The polygons with 2190_a habitat (dune slack ponds) are generated by selecting all watersurface polygons that overlap with dune habitat polygons (21xx) of the standardized habitat map.

For each of the other considered habitat types (31xx and rbbah) we select the watersurface polygons that overlap with the selected habitat type polygons of the standardized habitat map. We also select polygons of the standardized habitat map that contain standing water types but do not overlap with any watersurface polygon of the watersurface map.

The R-code for creating the watersurfaces_hab data source can be found in the n2khab-preprocessing repository.

References

Leyssen A., Scheers K., Smeekens V., Wils C., Packet J., De Knijf G. & Denys L. (2020). Watervlakken versie 1.1: polygonenkaart van stilstaand water in Vlaanderen. Uitgave 2020. Rapporten van het Instituut voor Natuur- en Bosonderzoek 2020 (40). Instituut voor Natuur en Bosonderzoek, Brussel. doi:10.21436/inbor.19088385 .

De Saeger, S., Guelinckx, R., Oosterlynck, P., De Bruyn, A., Debusschere, K., Dhaluin, P., Erens, R., Hendrickx, P., Hennebel, D., Jacobs, I., Kumpen, M., Op De Beeck, J., Spanhove, T., Tamsyn, W., Van Oost, F., Van Dam, G., Van Hove, M., Wils, C., Paelinckx, D. (2020). Biologische Waarderingskaart en Natura 2000 Habitatkaart, uitgave 2020. (Rapporten van het Instituut voor Natuur- en Bosonderzoek; Nr. 35). Instituut voor Natuur- en Bosonderzoek (INBO). doi:10.21436/inbor.18840851 .

See also

Examples

if (FALSE) {
# This example supposes that your working directory or a directory up to 10
# levels above has the 'n2khab_data' folder AND that the latest version of
# the 'watersurfaces_hab'
# data source is present in the default subdirectory.
# In all other cases, this example won't work but at least you can
# consider what to do.

wsh <- read_watersurfaces_hab()
wsh_polygons <- wsh$watersurfaces_polygons
wsh_types <- wsh$watersurfaces_types
wsh_polygons
wsh_types
}